Move JS-related types
