+++ tools/mc-js/src/main/java/io/spine/js/generate/GenerationTask.java
@@ -26,8 +26,8 @@
 
 package io.spine.js.generate;
 
-import io.spine.code.fs.js.Directory;
-import io.spine.code.fs.js.FileName;
+import io.spine.tools.js.fs.Directory;
+import io.spine.tools.js.fs.FileName;
 import io.spine.code.proto.FileSet;
 import io.spine.code.proto.ProtoBelongsToModule;
 import io.spine.code.proto.SourceFile;
