+++ tools/mc-js/src/main/java/io/spine/js/generate/field/parser/EnumFieldParser.java
@@ -28,7 +28,7 @@ package io.spine.js.generate.field.parser;
 
 import com.google.protobuf.Descriptors.EnumDescriptor;
 import com.google.protobuf.Descriptors.FieldDescriptor;
-import io.spine.code.gen.js.TypeName;
+import io.spine.tools.js.code.TypeName;
 import io.spine.js.generate.output.CodeLines;
 import io.spine.js.generate.output.snippet.VariableDeclaration;
