+++ tools/mc-js/src/main/java/io/spine/js/generate/index/KnownTypesMap.java
@@ -27,7 +27,7 @@
 package io.spine.js.generate.index;
 
 import com.google.common.collect.Maps;
-import io.spine.code.gen.js.TypeName;
+import io.spine.tools.js.code.TypeName;
 import io.spine.code.proto.FileSet;
 import io.spine.code.proto.TypeSet;
 import io.spine.js.generate.Snippet;
