+++ tools/mc-js/src/main/java/io/spine/js/generate/output/FileWriter.java
@@ -30,9 +30,9 @@ import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Charsets;
 import com.google.common.collect.ImmutableList;
 import com.google.protobuf.Descriptors.FileDescriptor;
-import io.spine.code.fs.js.Directory;
-import io.spine.code.fs.js.FileName;
-import io.spine.code.fs.js.LibraryFile;
+import io.spine.tools.js.fs.Directory;
+import io.spine.tools.js.fs.FileName;
+import io.spine.tools.js.fs.LibraryFile;
 
 import java.io.IOException;
 import java.nio.file.Files;
