+++ tools/mc-js/src/main/java/io/spine/js/generate/output/snippet/Import.java
@@ -26,7 +26,7 @@
 
 package io.spine.js.generate.output.snippet;
 
-import io.spine.code.fs.js.FileName;
+import io.spine.tools.js.fs.FileName;
 import io.spine.js.generate.output.CodeLine;
 
 import static com.google.common.base.Preconditions.checkNotNull;
