+++ tools/mc-js/src/main/java/io/spine/js/generate/output/snippet/VariableDeclaration.java
@@ -26,7 +26,7 @@
 
 package io.spine.js.generate.output.snippet;
 
-import io.spine.code.gen.js.TypeName;
+import io.spine.tools.js.code.TypeName;
 import io.spine.js.generate.output.CodeLine;
 
 import static com.google.common.base.Preconditions.checkNotNull;
