+++ tools/mc-js/src/main/java/io/spine/js/generate/parse/GeneratedParser.java
@@ -29,8 +29,8 @@ package io.spine.js.generate.parse;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.protobuf.Descriptors.Descriptor;
 import com.google.protobuf.Descriptors.FieldDescriptor;
-import io.spine.code.gen.js.MethodReference;
-import io.spine.code.gen.js.TypeName;
+import io.spine.tools.js.code.MethodReference;
+import io.spine.tools.js.code.TypeName;
 import io.spine.js.generate.Snippet;
 import io.spine.js.generate.field.FieldGenerator;
 import io.spine.js.generate.field.FieldGenerators;
