+++ tools/mc-js/src/main/java/io/spine/js/gradle/JsProtocConfigurationPlugin.java
@@ -27,7 +27,7 @@
 package io.spine.js.gradle;
 
 import com.google.protobuf.gradle.ExecutableLocator;
-import io.spine.code.fs.js.DefaultJsProject;
+import io.spine.tools.js.fs.DefaultJsProject;
 import io.spine.tools.gradle.ProtocConfigurationPlugin;
 import org.gradle.api.NamedDomainObjectContainer;
 import org.gradle.api.Project;
