+++ tools/mc-js/src/test/java/io/spine/js/generate/field/FieldGeneratorTest.java
@@ -27,7 +27,7 @@
 package io.spine.js.generate.field;
 
 import com.google.protobuf.Descriptors.FieldDescriptor;
-import io.spine.code.gen.js.FieldName;
+import io.spine.tools.js.code.FieldName;
 import io.spine.js.generate.output.CodeLines;
 import io.spine.type.TypeUrl;
 import org.junit.jupiter.api.BeforeEach;
