+++ tools/mc-js/src/test/java/io/spine/js/generate/given/GivenMethod.java
@@ -27,8 +27,8 @@
 package io.spine.js.generate.given;
 
 import com.google.protobuf.Any;
-import io.spine.code.gen.js.MethodReference;
-import io.spine.code.gen.js.TypeName;
+import io.spine.tools.js.code.MethodReference;
+import io.spine.tools.js.code.TypeName;
 
 public final class GivenMethod {
