+++ tools/mc-js/src/test/java/io/spine/js/generate/given/TestGenerationTask.java
@@ -26,7 +26,7 @@
 
 package io.spine.js.generate.given;
 
-import io.spine.code.fs.js.Directory;
+import io.spine.tools.js.fs.Directory;
 import io.spine.code.proto.FileSet;
 import io.spine.js.generate.GenerationTask;
