+++ tools/mc-js/src/test/java/io/spine/js/generate/imports/ResolveImportsTest.java
@@ -28,7 +28,7 @@ package io.spine.js.generate.imports;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.truth.IterableSubject;
-import io.spine.code.fs.js.Directory;
+import io.spine.tools.js.fs.Directory;
 import io.spine.js.generate.given.GivenProject;
 import io.spine.tools.fs.DirectoryPattern;
 import io.spine.tools.fs.ExternalModule;
