+++ tools/mc-js/src/test/java/io/spine/js/generate/index/GenerateIndexFileTest.java
@@ -27,8 +27,8 @@
 package io.spine.js.generate.index;
 
 import com.google.protobuf.Descriptors.FileDescriptor;
-import io.spine.code.fs.js.Directory;
-import io.spine.code.fs.js.FileName;
+import io.spine.tools.js.fs.Directory;
+import io.spine.tools.js.fs.FileName;
 import io.spine.code.proto.FileSet;
 import io.spine.js.generate.given.GivenProject;
 import io.spine.js.generate.output.CodeLines;
@@ -37,7 +37,7 @@ import org.junit.jupiter.api.Test;
 
 import java.nio.file.Path;
 
-import static io.spine.code.fs.js.LibraryFile.INDEX;
+import static io.spine.tools.js.fs.LibraryFile.INDEX;
 import static io.spine.js.generate.given.Generators.assertContains;
 import static java.nio.file.Files.exists;
 import static org.junit.jupiter.api.Assertions.assertTrue;
