+++ tools/mc-js/src/test/java/io/spine/js/generate/parse/GenerateKnownTypeParsersTest.java
@@ -30,9 +30,9 @@ import com.google.common.testing.NullPointerTester;
 import com.google.protobuf.Any;
 import com.google.protobuf.Descriptors.Descriptor;
 import com.google.protobuf.Descriptors.FileDescriptor;
-import io.spine.code.fs.js.Directory;
-import io.spine.code.fs.js.FileName;
-import io.spine.code.gen.js.TypeName;
+import io.spine.tools.js.fs.Directory;
+import io.spine.tools.js.fs.FileName;
+import io.spine.tools.js.code.TypeName;
 import io.spine.code.proto.FileDescriptors;
 import io.spine.code.proto.FileSet;
 import io.spine.code.proto.TypeSet;
