+++ tools/mc-js/src/test/java/io/spine/js/gradle/ExtensionTest.java
@@ -26,8 +26,8 @@
 
 package io.spine.js.gradle;
 
-import io.spine.code.fs.js.DefaultJsProject;
-import io.spine.code.fs.js.Directory;
+import io.spine.tools.js.fs.DefaultJsProject;
+import io.spine.tools.js.fs.Directory;
 import io.spine.tools.fs.ExternalModule;
 import org.gradle.api.Project;
 import org.gradle.api.plugins.PluginManager;
