+++ tools/tool-base/src/main/java/io/spine/tools/js/code/MethodReference.java
@@ -24,16 +24,46 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+package io.spine.tools.js.code;
+
+import io.spine.value.StringTypeValue;
+
+import static java.lang.String.format;
+
 /**
- * This package contains tools for working with the structure of a JavaScript project.
+ * A reference to a method of a Protobuf type.
+ *
+ * <p>The reference allows to identify a method in code
+ * since it includes a type name and a method name.
  */
+public final class MethodReference extends StringTypeValue {
 
-@Internal
-@CheckReturnValue
-@ParametersAreNonnullByDefault
-package io.spine.code.fs.js;
+    private static final long serialVersionUID = 0L;
 
-import com.google.errorprone.annotations.CheckReturnValue;
-import io.spine.annotation.Internal;
+    private MethodReference(String value) {
+        super(value);
+    }
 
-import javax.annotation.ParametersAreNonnullByDefault;
+    /**
+     * Obtains the reference to the instance method of the specified type.
+     */
+    public static MethodReference onType(TypeName typeName, String methodName) {
+        String value = format("%s.%s", typeName, methodName);
+        return new MethodReference(value);
+    }
+
+    /**
+     * Obtains the reference to the static method of the specified type.
+     */
+    public static MethodReference onPrototype(TypeName typeName, String methodName) {
+        String value = format("%s.prototype.%s", typeName, methodName);
+        return new MethodReference(value);
+    }
+
+    /**
+     * Obtains the reference to the constructor of the specified type.
+     */
+    public static MethodReference constructor(TypeName typeName) {
+        return new MethodReference(typeName.value());
+    }
+}
