+++ tools/tool-base/src/main/java/io/spine/tools/js/fs/DefaultJsProject.java
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.code.fs.js;
+package io.spine.tools.js.fs;
 
 import io.spine.tools.fs.DefaultProject;
