+++ tools/tool-base/src/main/java/io/spine/tools/proto/fs/Directory.java
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.code.fs.js;
+package io.spine.tools.proto.fs;
 
 import io.spine.code.AbstractDirectory;
 import io.spine.code.SourceCodeDirectory;
@@ -34,48 +34,32 @@ import java.nio.file.Path;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 /**
- * A folder with JavaScript source files.
+ * A proto source code directory.
  */
 public final class Directory extends SourceCodeDirectory {
 
-    private static final String ROOT_NAME = "js";
+    @SuppressWarnings("DuplicateStringLiteralInspection") // Same name for different directories.
+    private static final String ROOT_NAME = "proto";
 
     private Directory(Path path) {
         super(path);
     }
 
     /**
-     * Creates a new instance at the specified location.
+     * Creates a new instance.
      */
-    public static Directory at(Path path) {
+    static Directory at(Path path) {
         checkNotNull(path);
         return new Directory(path);
     }
 
     /**
-     * Creates an instance of the root directory named {@code "js"}.
+     * Creates an instance of the root directory named {@code "proto"}.
      */
-    static Directory rootIn(AbstractDirectory parent) {
+    public static Directory rootIn(AbstractDirectory parent) {
         checkNotNull(parent);
         Path path = parent.path()
                           .resolve(ROOT_NAME);
         return at(path);
-    }
-
-    /**
-     * Obtains the source code path for the passed file name.
-     */
-    public Path resolve(FileName fileName) {
-        checkNotNull(fileName);
-        Path result = path().resolve(fileName.value());
-        return result;
-    }
-
-    /**
-     * Obtains the source code path for the passed library file.
-     */
-    public Path resolve(LibraryFile libraryFile) {
-        checkNotNull(libraryFile);
-        return resolve(libraryFile.fileName());
     }
 }
